Update two_sum.hpp
Switch from hash map to sparse array, to improve performance (by not needing to hash or deal with hash collisions) at the expense of memory (for empty slots.)  Using pair to store boolean flag marking used slots in array, since index values cover entire range of integers.  But, this implementation will only work for a range of integers in the input array that is non-negative.